A new approach
I shall try a new approach.
Set everything in gbuffer including albedo roughness etc.
Let go of ibl approach as a fall back, i will just do a pbr in
the lightening pass.

The lightening pass would be done in view space.
This means output values from gbuffer in viewspace.
From depth info we can construct frag position in view space,
which is okay.

The normal should be expressed in view space as well.
It can be stored in depth buffer, in normalized form.
Its length is also required unlike frag pos.
because the ray from camera to normal can not be deduced in the
lightening pass.

Camera position also needs to be expressed in view space
but that is doable during the lightening phase as well.