Attempting to implement conetracing now.
Here is the current status, the ibl portion works, but the lightening
does not look so good. I will try to implement the cone tracing now.
In order to do that I need to capture the output of the lightening
to a texture then use it in the cone tracing shader.
So a fbo is needed for lightening.