Several ideas for accelarating everything in defered
In a defered set up we can do the following.

Cherry pick pixels for tracing using roughness
and metallic map as a heuristic.
generate ray towards light source and objects
of interest.
So the algorithm is mark objects of interest
with using roughness and metallic maps in
gbuffer then sample rays towards them and the
light source.